Isolated piece isolation into it's own class.

+++ src/engine/inc/move/move.hpp
@@ -295,7 +295,7 @@ struct PrioritizedMove {
         priority(0),
         check(0) {};
 
-    PrioritizedMove(PackedMove move, int _priority) :
+    PrioritizedMove(PackedMove move, u16 _priority) :
         move(move),
         priority(_priority),
         check(0) {};
